할일 제목 비밀번호 필수값 처리 메니저는 이메일형식으로만 받도록 처리

+++ src/main/java/com/sparta/todo/service/TodoService.java
@@ -1,9 +1,9 @@
 package com.sparta.todo.service;
 
+import com.sparta.todo.apiResponse.ApiResponse;
 import com.sparta.todo.dto.TodoRequestDto;
 import com.sparta.todo.dto.TodoResponseDto;
 import com.sparta.todo.entity.Todo;
-import com.sparta.todo.exception.HttpException;
 import com.sparta.todo.repository.TodoRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -19,45 +19,56 @@ public class TodoService {
         this.todoRepository = todoRepository;
     }
 
-    public TodoResponseDto createTodo(TodoRequestDto todoRequestDto){
+    public ApiResponse<TodoResponseDto> createTodo(TodoRequestDto todoRequestDto){
         Todo todo = new Todo(todoRequestDto);
         Todo saveTodo = todoRepository.save(todo);
         TodoResponseDto todoResponseDto = new TodoResponseDto(todo);
-
-        return todoResponseDto;
+        ApiResponse<TodoResponseDto> apiResponse = new ApiResponse<>(HttpStatus.OK,"할일 생성",todoResponseDto);
+        return apiResponse;
     }
 
-    public TodoResponseDto inquireTodo(Long id){
+    public ApiResponse<TodoResponseDto> inquireTodo(Long id){
 
         Todo todo = findByTodo(id);
         TodoResponseDto todoResponseDto = new TodoResponseDto(todo);
-        return todoResponseDto;
+        ApiResponse<TodoResponseDto> apiResponse = new ApiResponse(HttpStatus.OK,"조회 성공",todoResponseDto);
+        return apiResponse;
     }
-    public List<TodoResponseDto> getTodo(){
-        return todoRepository.findAllByOrderByModifiedAtDesc().stream().map(TodoResponseDto::new).toList();
+    public ApiResponse<List> getTodo(){
+        List<TodoResponseDto> todoList = todoRepository.findAllByOrderByModifiedAtDesc().stream().map(TodoResponseDto::new).toList();
+        ApiResponse<List> apiResponse = new ApiResponse<>(HttpStatus.OK,"조회 성공",todoList);
+        return apiResponse;
     }
 
     @Transactional
-    public Long updateTodo(TodoRequestDto todoRequestDto,Long id){
+    public ApiResponse<Long> updateTodo(TodoRequestDto todoRequestDto,Long id){
         Todo todo = findByTodo(id);
+        ApiResponse<Long> apiResponse;
+        if(todo == null)
+            return new ApiResponse(HttpStatus.NOT_FOUND,id+"는 존재하지 않는 ID입니다.",id);
         String password = validationPassword(todoRequestDto,id);
         if(!password.equals("0")){
             todo.update(todoRequestDto);
+            apiResponse = new ApiResponse<Long>(HttpStatus.OK,"수정 성공",id);
         }else{
-            throw new HttpException("패스워드가 틀렸습니다.", HttpStatus.FORBIDDEN);
+            apiResponse = new ApiResponse(HttpStatus.FORBIDDEN,"수정 실패",id);
         }
 
-        return id;
+        return apiResponse;
     }
 
-    public Long deleteTodo(Long id,String password){
+    public ApiResponse<Long> deleteTodo(Long id,String password){
         Todo todo = findByTodo(id);
+        ApiResponse<Long> apiResponse;
+        if(todo == null)
+            return new ApiResponse(HttpStatus.NOT_FOUND,id+"는 존재하지 않는 ID입니다.",id);
         if(todo.getPassword().equals(password)){
             todoRepository.delete(todo);
+            apiResponse = new ApiResponse<Long>(HttpStatus.OK,"삭제 성공",id);
+        }else{
+            apiResponse = new ApiResponse<Long>(HttpStatus.FORBIDDEN,"삭제 실패",id);
         }
-
-
-        return id;
+        return apiResponse;
     }
     public String validationPassword(TodoRequestDto todoRequestDto,Long id){
         Todo todo = findByTodo(id);
@@ -68,7 +79,7 @@ public class TodoService {
     }
 
     public Todo findByTodo(Long id){
-        return todoRepository.findById(id).orElseThrow(()-> new HttpException("없는 id입니다.",HttpStatus.NOT_FOUND));
+        return todoRepository.findById(id).orElseThrow(()-> null);
     }
